Update 01_find_overlap.Rmd
For the life of me i couldnt find the /food_tree_gits folder for the first file. Also, in putting together the abc file, i never got 8 columns. Maybe there was an update to read.delim or something and row.names are not longer read in? Or maybe im doing something wrong. Seems like it works the way it should with 5:7 (line 95).